Stars-pm : force float in kernel
Constant has to be "f" suffixed to be considered as float, else on recent GPU they would be compiled to double operations with final conversion to float : costly !

+++ examples/P4A/Stars-pm/include/varglob.h
@@ -7,11 +7,11 @@
 #endif
 
 #define NPART (NP*NP*NP)
-#define LBOX 6.
-#define G 1.
-#define TMAX 7.5
-#define DX (LBOX/NP)
-#define DT (5*1e-2)
+#define LBOX 6.f
+#define G 1.f
+#define TMAX 7.5f
+#define DX ((float)LBOX/NP)
+#define DT ((float)5.f*1e-2f)
 #define NPBLOCK 128
 #define BLOCK_SIZE 2
 #define MODT 1000
@@ -19,7 +19,7 @@
 #define CUERR() printf("\n %s \n",cudaGetErrorString(cudaGetLastError()))
 
 #ifndef M_PI
-#define M_PI (3.14159265358979323846)
+#define M_PI (3.14159265358979323846f)
 #endif
